* signal.c (rb_sigaltstack_size): cast sysconf() return value   explicitly. Fix compile error on Mac OS X.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@38415 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ signal.c
@@ -459,7 +459,7 @@ int rb_sigaltstack_size(void)
 #if defined(HAVE_SYSCONF) && defined(_SC_PAGE_SIZE)
     {
 	int pagesize;
-	pagesize = sysconf(_SC_PAGE_SIZE);
+	pagesize = (int)sysconf(_SC_PAGE_SIZE);
 	if (size < pagesize)
 	    size = pagesize;
     }
